Share more code between loki and prometheus query modellers
