+++ public/app/plugins/datasource/loki/querybuilder/LokiQueryModeller.ts
@@ -1,9 +1,9 @@
 import { LokiAndPromQueryModellerBase } from '../../prometheus/querybuilder/shared/LokiAndPromQueryModellerBase';
 import { QueryBuilderLabelFilter } from '../../prometheus/querybuilder/shared/types';
 import { getOperationDefintions } from './operations';
-import { LokiOperationId, LokiQueryPattern, LokiVisualQuery, LokiVisualQueryOperationCategory } from './types';
+import { LokiOperationId, LokiQueryPattern, LokiVisualQueryOperationCategory } from './types';
 
-export class LokiQueryModeller extends LokiAndPromQueryModellerBase<LokiVisualQuery> {
+export class LokiQueryModeller extends LokiAndPromQueryModellerBase {
   constructor() {
     super(getOperationDefintions);
 
@@ -25,31 +25,6 @@ export class LokiQueryModeller extends LokiAndPromQueryModellerBase<LokiVisualQu
     return super.renderLabels(labels);
   }
 
-  renderQuery(query: LokiVisualQuery, nested?: boolean) {
-    let queryString = `${this.renderLabels(query.labels)}`;
-    queryString = this.renderOperations(queryString, query.operations);
-
-    if (!nested && this.hasBinaryOp(query) && Boolean(query.binaryQueries?.length)) {
-      queryString = `(${queryString})`;
-    }
-
-    queryString = this.renderBinaryQueries(queryString, query.binaryQueries);
-
-    if (nested && (this.hasBinaryOp(query) || Boolean(query.binaryQueries?.length))) {
-      queryString = `(${queryString})`;
-    }
-
-    return queryString;
-  }
-
-  hasBinaryOp(query: LokiVisualQuery): boolean {
-    return (
-      query.operations.find((op) => {
-        const def = this.getOperationDef(op.id);
-        return def?.category === LokiVisualQueryOperationCategory.BinaryOps;
-      }) !== undefined
-    );
-  }
   getQueryPatterns(): LokiQueryPattern[] {
     return [
       {
