+++ public/app/plugins/datasource/prometheus/querybuilder/PromQueryModeller.ts
@@ -2,9 +2,9 @@ import { FUNCTIONS } from '../promql';
 import { getAggregationOperations } from './aggregations';
 import { getOperationDefinitions } from './operations';
 import { LokiAndPromQueryModellerBase } from './shared/LokiAndPromQueryModellerBase';
-import { PromQueryPattern, PromVisualQuery, PromVisualQueryOperationCategory } from './types';
+import { PromQueryPattern, PromVisualQueryOperationCategory } from './types';
 
-export class PromQueryModeller extends LokiAndPromQueryModellerBase<PromVisualQuery> {
+export class PromQueryModeller extends LokiAndPromQueryModellerBase {
   constructor() {
     super(() => {
       const allOperations = [...getOperationDefinitions(), ...getAggregationOperations()];
@@ -25,32 +25,6 @@ export class PromQueryModeller extends LokiAndPromQueryModellerBase<PromVisualQu
       PromVisualQueryOperationCategory.Trigonometric,
       PromVisualQueryOperationCategory.Time,
     ]);
-  }
-
-  renderQuery(query: PromVisualQuery, nested?: boolean) {
-    let queryString = `${query.metric}${this.renderLabels(query.labels)}`;
-    queryString = this.renderOperations(queryString, query.operations);
-
-    if (!nested && this.hasBinaryOp(query) && Boolean(query.binaryQueries?.length)) {
-      queryString = `(${queryString})`;
-    }
-
-    queryString = this.renderBinaryQueries(queryString, query.binaryQueries);
-
-    if (nested && (this.hasBinaryOp(query) || Boolean(query.binaryQueries?.length))) {
-      queryString = `(${queryString})`;
-    }
-
-    return queryString;
-  }
-
-  hasBinaryOp(query: PromVisualQuery): boolean {
-    return (
-      query.operations.find((op) => {
-        const def = this.getOperationDef(op.id);
-        return def?.category === PromVisualQueryOperationCategory.BinaryOps;
-      }) !== undefined
-    );
   }
 
   getQueryPatterns(): PromQueryPattern[] {
