+++ public/app/plugins/datasource/prometheus/querybuilder/shared/LokiAndPromQueryModellerBase.ts
@@ -1,11 +1,6 @@
 import { Registry } from '@grafana/data';
-import {
-  QueryBuilderLabelFilter,
-  QueryBuilderOperation,
-  QueryBuilderOperationDef,
-  QueryWithOperations,
-  VisualQueryModeller,
-} from './types';
+import { PromVisualQueryOperationCategory } from '../types';
+import { QueryBuilderLabelFilter, QueryBuilderOperation, QueryBuilderOperationDef, VisualQueryModeller } from './types';
 
 export interface VisualQueryBinary<T> {
   operator: string;
@@ -14,7 +9,14 @@ export interface VisualQueryBinary<T> {
   query: T;
 }
 
-export abstract class LokiAndPromQueryModellerBase<T extends QueryWithOperations> implements VisualQueryModeller {
+export interface PromLokiVisualQuery {
+  metric?: string;
+  labels: QueryBuilderLabelFilter[];
+  operations: QueryBuilderOperation[];
+  binaryQueries?: Array<VisualQueryBinary<PromLokiVisualQuery>>;
+}
+
+export abstract class LokiAndPromQueryModellerBase implements VisualQueryModeller {
   protected operationsRegisty: Registry<QueryBuilderOperationDef>;
   private categories: string[] = [];
 
@@ -54,7 +56,7 @@ export abstract class LokiAndPromQueryModellerBase<T extends QueryWithOperations
     return queryString;
   }
 
-  renderBinaryQueries(queryString: string, binaryQueries?: Array<VisualQueryBinary<T>>) {
+  renderBinaryQueries(queryString: string, binaryQueries?: Array<VisualQueryBinary<PromLokiVisualQuery>>) {
     if (binaryQueries) {
       for (const binQuery of binaryQueries) {
         queryString = `${this.renderBinaryQuery(queryString, binQuery)}`;
@@ -63,7 +65,7 @@ export abstract class LokiAndPromQueryModellerBase<T extends QueryWithOperations
     return queryString;
   }
 
-  private renderBinaryQuery(leftOperand: string, binaryQuery: VisualQueryBinary<T>) {
+  private renderBinaryQuery(leftOperand: string, binaryQuery: VisualQueryBinary<PromLokiVisualQuery>) {
     let result = leftOperand + ` ${binaryQuery.operator} `;
 
     if (binaryQuery.vectorMatches) {
@@ -90,5 +92,29 @@ export abstract class LokiAndPromQueryModellerBase<T extends QueryWithOperations
     return expr + `}`;
   }
 
-  abstract renderQuery(query: T, nested?: boolean): string;
+  renderQuery(query: PromLokiVisualQuery, nested?: boolean) {
+    let queryString = `${query.metric ?? ''}${this.renderLabels(query.labels)}`;
+    queryString = this.renderOperations(queryString, query.operations);
+
+    if (!nested && this.hasBinaryOp(query) && Boolean(query.binaryQueries?.length)) {
+      queryString = `(${queryString})`;
+    }
+
+    queryString = this.renderBinaryQueries(queryString, query.binaryQueries);
+
+    if (nested && (this.hasBinaryOp(query) || Boolean(query.binaryQueries?.length))) {
+      queryString = `(${queryString})`;
+    }
+
+    return queryString;
+  }
+
+  hasBinaryOp(query: PromLokiVisualQuery): boolean {
+    return (
+      query.operations.find((op) => {
+        const def = this.getOperationDef(op.id);
+        return def?.category === PromVisualQueryOperationCategory.BinaryOps;
+      }) !== undefined
+    );
+  }
 }
